fix: ethers v5.4.1
EthersJS 5.4.0 releases support for EIP1559 new transaction format

This library was previously written to be tightly coupled to the keys in the transaction request (why? not sure) - this was causing a malformed transaction request with ethers 5.4.1 as it was putting in unexpected object keys

This has been fixed in this release - update if you find that your transactions seem to be getting sent but not showing up as confirmed.